zero branch 머지 전 커밋

+++ src/main/java/com/bit/kodari/service/TradeService.java
@@ -205,8 +205,28 @@ public class TradeService {
     // 거래내역 삭제 : status 수정
     @Transactional // Trancaction 기능 : 데이터 생성,수정,삭제와같은 데이터를 작업하는 일이 여러 과정을 한번에 수행 항 때 수행을 끝마쳐야 저장, 오류나면 Rollback 해서 안전성을 부여.
     public void deleteTrade(TradeDto.PatchStatusReq patchStatusReq) throws BaseException{
-
+        // 이미 삭제된 거래내역 validation
+        String status = tradeRepository.getStatusByTradeIdx(patchStatusReq.getTradeIdx());
+        if(status.equals("inactive")){
+            throw new BaseException(BaseResponseStatus.ALREADY_DELETED_TRADE); //
+        }
+        // 거래내역 삭제 요청
         int result = tradeRepository.deleteTrade(patchStatusReq);
+        if (result == 0) {// result값이 0이면 과정이 실패한 것이므로 에러 메서지를 보냅니다.
+            throw new BaseException(BaseResponseStatus.REQUEST_ERROR);
+        }
+    }
+
+    // 거래내역 삭제 : 전체삭제
+    @Transactional
+    public void deleteAllTradeByUserIdx(int userIdx) throws BaseException{
+        // 이미 삭제된 거래내역 validation
+        /*String status = tradeRepository.getStatusByTradeIdx(patchStatusReq.getTradeIdx());
+        if(status.equals("inactive")){
+            throw new BaseException(BaseResponseStatus.ALREADY_DELETED_TRADE); //
+        }*/
+        // 거래내역 삭제 요청
+        int result = tradeRepository.deleteAllTradeByUserIdx(userIdx);
         if (result == 0) {// result값이 0이면 과정이 실패한 것이므로 에러 메서지를 보냅니다.
             throw new BaseException(BaseResponseStatus.REQUEST_ERROR);
         }
